Update download locations
The current code of Installer does not follow recursive redirects during download the modx package. It follows just one redirect. So the download url has to be the url one redirect before the S3 bucket. This is the case with my commit.

I also have fixed some wrong tab indents.